Initial migration debugs

+++ src/Controller/TaskController.php
@@ -3,7 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Task;
-use AppBundle\Form\TaskType;
+use App\Form\TaskType;
 use App\Repository\TaskRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -11,16 +11,18 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
+
+#[Route('/tasks')]
 class TaskController extends AbstractController
 {
-    #[Route('/tasks', name: 'task_list', methods: ['GET'])]
-    public function listAction(TaskRepository $taskRepository) : Response
+    #[Route('', name: 'task_list', methods: ['GET'])]
+    public function list(TaskRepository $taskRepository) : Response
     {
         return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
     }
 
-    #[Route('/tasks/create', name: 'task_create', methods: ['GET', 'POST'])]
-     public function createAction(Request $request, TaskRepository $taskRepository) : Response
+    #[Route('/create', name: 'task_create', methods: ['GET', 'POST'])]
+     public function create(Request $request, TaskRepository $taskRepository) : Response
     {
         /** @var \App\Entity\User $user */
         $user = $this->getUser();
@@ -41,8 +43,8 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form]);
     }
 
-    #[Route('/tasks/{id}/edit', name: 'task_edit', methods: ['GET', 'POST'])]
-     public function editAction(#[MapEntity(id:'id')]Task $task, Request $request, TaskRepository $taskRepository) : Response
+    #[Route('/{id}/edit', name: 'task_edit', methods: ['GET', 'POST'])]
+     public function edit(#[MapEntity(id:'id')]Task $task, Request $request, TaskRepository $taskRepository) : Response
     {
         $form = $this->createForm(TaskType::class, $task);
 
@@ -62,8 +64,8 @@ class TaskController extends AbstractController
         ]);
     }
 
-    #[Route('/tasks/{id}/toggle', name: 'task_toggle', methods: ['POST'])]
-    public function toggleTaskAction(#[MapEntity(id:'id')]Task $task, TaskRepository $taskRepository) : Response
+    #[Route('/{id}/toggle', name: 'task_toggle', methods: ['POST'])]
+    public function toggleTask(#[MapEntity(id:'id')]Task $task, TaskRepository $taskRepository) : Response
     {
         $task->toggle(!$task->isIsDone());
         $taskRepository->save($task, true);
@@ -73,13 +75,14 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list', [], Response::HTTP_SEE_OTHER);
     }
 
-    #[Route('/tasks/{id}/delete', name: 'task_delete', methods: ['POST'])]
-    public function deleteTaskAction(#[MapEntity(id:'id')]Task $task, TaskRepository $taskRepository) : Response
+    #[Route('/{id}/delete', name: 'task_delete', methods: ['POST'])]
+    public function deleteTask(#[MapEntity(id:'id')]Task $task, TaskRepository $taskRepository, Request $request) : Response
     {
-        $taskRepository->remove($task, true);
-
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
-
+        if ($this->isCsrfTokenValid(sprintf('delete%s', $task->getId()), $request->request->get('_token'))) {
+            $taskRepository->remove($task, true);
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        }
+        
         return $this->redirectToRoute('task_list', [], Response::HTTP_SEE_OTHER);
     }
 }
